Update examples and improve consistency
- Centralized `configure_target_options` usage across examples.
- Enhanced type consistency and precision in `GameOfLife` and `Engine` with `unsigned long` and `[[maybe_unused]]` annotations.
- Refined random number generation using new `randf` and `randi` methods.
- Adjusted `Renderer` to use `SliderInt` for frequency control.
- Introduced `life` example with conditional build support.

+++ examples/gol/include/game_of_life.hpp
@@ -20,9 +20,9 @@ namespace gol {
         void new_world();
 
     private:
-        static constexpr int GRID_WIDTH = 250;
-        static constexpr int GRID_HEIGHT = 250;
-        int count_neighbors(int x, int y);
+        static constexpr unsigned long GRID_WIDTH = 250;
+        static constexpr unsigned long GRID_HEIGHT = 250;
+        unsigned long count_neighbors(unsigned long x, unsigned long y);
 
         std::vector<std::vector<bool> > m_grid;
         std::vector<std::vector<bool> > m_next_grid;
